+++ services/ledger-writer/app/models/trade.py
@@ -1,7 +1,7 @@
 import uuid
 
 from app.db.base import Base
-from sqlalchemy import Column, Float, ForeignKey, String
+from sqlalchemy import Column, Float, ForeignKey, Numeric, String
 from sqlalchemy.dialects.postgresql import UUID
 
 
@@ -11,6 +11,6 @@ class Trade(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     order_id = Column(UUID(as_uuid=True), ForeignKey("orders.id"), nullable=False)
     symbol = Column(String, nullable=False)
-    price = Column(Float, nullable=False)
-    quantity = Column(Float, nullable=False)
+    price = Column(Numeric, nullable=False)
+    quantity = Column(Numeric, nullable=False)
     timestamp = Column(Float, nullable=False)
